fix:move dump db to migrations folder fix: create classes folder. Move in this folder classes которые обрабатывают машруты.
