+++ classes/Admin.js
@@ -1,0 +1,17 @@
+const sql = require("../models/postgres.js");
+
+class Admin {
+  async getArtists(req, res, next) {
+    try {
+      const users = await sql`
+        select * from artists
+        `;
+
+      res.json(users);
+    } catch (e) {
+      console.log(e);
+      next(e);
+    }
+  }
+}
+module.exports = new Admin();
