+++ middlewares/ErrorMiddleware.js
@@ -1,0 +1,11 @@
+const ApiError = require("../exceptions/ApiError.js");
+
+module.exports = function (err, req, res, next) {
+  console.log(err);
+  if (err instanceof ApiError) {
+    return res
+      .status(err.status)
+      .json({ message: err.message, errors: err.errors });
+  }
+  return res.status(500).json({ message: "Непредвиденная ошибка" });
+};
