+++ router/index.js
@@ -1,22 +1,22 @@
 const Router = require("express").Router;
-const userClass = require("../controllers/user-controller");
+const Auth = require("../classes/Auth.js");
 const router = new Router();
 const { body } = require("express-validator");
-const authMiddleware = require("../middlewares/auth-middleware");
-const authMiddlewareAdmin = require("../middlewares/auth-middleware-admin");
-const getArtists = require("../controllers/getArtists.js");
-const getAboutMe = require("../controllers/getAboutMe.js");
+const authMiddleware = require("../middlewares/AuthMiddleware.js");
+const authMiddlewareAdmin = require("../middlewares/AuthMiddlewareAdmin.js");
+const Artist = require("./../classes/Artist.js");
+const Admin = require("./../classes/Admin.js");
 
 router.post(
   "/signup",
   body("email").isEmail(),
   body("password").isLength({ min: 3, max: 32 }),
-  userClass.registration
+  Auth.registration
 );
-router.post("/login", userClass.login);
-router.post("/logout", userClass.logout);
-router.get("/refresh", userClass.refresh);
-router.get("/about-me", authMiddleware, getAboutMe);
-router.get("/get-artists", authMiddlewareAdmin, getArtists);
+router.post("/login", Auth.login);
+router.post("/logout", Auth.logout);
+router.get("/refresh", Auth.refresh);
+router.get("/about-me", authMiddleware, Artist.getAboutMe);
+router.get("/get-artists", authMiddlewareAdmin, Admin.getArtists);
 
 module.exports = router;
